Made it so the buttons are always centered, as before it was too left-heavy or right-heavy depending on whether the row had long words or short words. In order to make this possible, I changed the way the tapping on words works during translations: now the buttons have a fixed position and return to it when the button is clicked at the top, rather than every button constantly readjusting its position based on the position of other buttons. Added frames to the buttons at the bottom so when the button is at the top, there is a gray frame that shows where it would return (its original position).
Added an Argentinian and an Australian flag to the launch screen to replace the buttons that said "ARG" and "AUS." Also added an Italian and a Brazilian flag, both of which are currently not clickable. Also added the current flag to the learning screen at the top, next to where it says the name of the app.

Fixed bug where the current audio clip would be deleted after an incorrect response, which would lead to a crash later on.